requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
pillow==8.3.1
tensorflow==2.6.0
numpy==1.19.5
werkzeug==2.0.1
matplotlib==3.4.3
pandas==1.3.3
protobuf==3.20.0
keras==2.6.0
flatbuffers==1.12.0
h5py==3.1.0
